backend/requirements.txt
django
djangorestframework
psycopg2
django-cors-headers
pillow
requests